fix: policy

+++ src/domains/policy/policy.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { ConflictException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { BaseResponse } from 'shared/generics/base.response';
@@ -18,6 +18,8 @@ export class PolicyService {
 
   async create(body: PolicyCreateREQ) {
     this.logger.log(`create policy: ${JSON.stringify(body)}`);
+    const policy = await this.policyModel.findOne({ name: body.name });
+    if (policy) throw new ConflictException(`Chính sách: ${body.name} đã tồn tại!`);
     const newPolicy = await this.policyModel.create(body);
     return BaseResponse.withMessage(toDocModel(newPolicy), 'Tạo chính sách thành công!');
   }
@@ -25,26 +27,24 @@ export class PolicyService {
   async findAllByObject(query: PolicyFindAllByObjectREQ) {
     this.logger.log(`find all policy by object: ${JSON.stringify(query)}`);
     const conditions = PolicyFindAllByObjectREQ.toQueryCondition(query);
-    const policies = await this.policyModel.find(conditions, { _id: 1, name: 1, content: 1 }, { lean: true });
-    const data = {};
-    policies.forEach((policy) => {
-      if (data.hasOwnProperty(policy.name)) {
-        data[policy.name].content += `\n${policy.content}`;
-      } else {
-        data[policy.name] = policy;
-      }
-    });
-    return BaseResponse.withMessage(Object.values(data), 'Lấy danh sách chính sách thành công!');
+    const policies = await this.policyModel
+      .find(conditions, { _id: 1, name: 1, content: 1 }, { lean: true })
+      .sort({ createdAt: -1 });
+    return BaseResponse.withMessage(policies, 'Lấy danh sách chính sách thành công!');
   }
 
   async update(id: string, body: PolicyUpdateREQ) {
     this.logger.log(`update policy: ${id}, ${JSON.stringify(body)}`);
+    const policy = await this.policyModel.findById(id);
+    if (!policy) throw new NotFoundException('Không tìm thấy chính sách!');
     const updatedPolicy = await this.policyModel.findByIdAndUpdate({ _id: id }, { ...body }, { new: true, lean: true });
     return BaseResponse.withMessage(updatedPolicy, 'Cập nhật chính sách thành công!');
   }
 
   async delete(id: string) {
     this.logger.log(`delete policy: ${id}`);
+    const policy = await this.policyModel.findById(id);
+    if (!policy) throw new NotFoundException('Không tìm thấy chính sách!');
     await this.policyModel.findByIdAndDelete(id);
     return BaseResponse.withMessage({}, 'Xóa chính sách thành công!');
   }
